feat: strictnullchecks  by default
Adding the strictnullchecks as a default flag to projects. Although the
output is generated slightly differently, it should be backwards
compatible.

The only difference I found was that we were adding both
`.optional()` and `.Or(Undefined)` to optional properties, which
according to the runtypes documentation will produce the same thing.